Implement StringUtilityTest with phpunit framework instead of deprecated simpletest framework
